feat(be): move group guards into roles module
