+++ backend/apps/client/src/contest/contest.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common'
 import { RolesModule } from '@libs/auth'
-import { GroupModule } from '@client/group/group.module'
 import {
   ContestAdminController,
   ContestPublicizingRequestAdminController,
@@ -11,7 +10,7 @@ import { ContestController, GroupContestController } from './contest.controller'
 import { ContestService } from './contest.service'
 
 @Module({
-  imports: [RolesModule, GroupModule],
+  imports: [RolesModule],
   controllers: [
     ContestController,
     GroupContestController,
