+++ backend/apps/client/src/notice/notice.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common'
 import { RolesModule } from '@libs/auth'
-import { GroupModule } from '@client/group/group.module'
 import {
   NoticeAdminController,
   GroupNoticeAdminController
@@ -9,7 +8,7 @@ import { NoticeController, GroupNoticeController } from './notice.controller'
 import { NoticeService } from './notice.service'
 
 @Module({
-  imports: [RolesModule, GroupModule],
+  imports: [RolesModule],
   controllers: [
     NoticeController,
     GroupNoticeController,
