+++ backend/apps/client/src/workbook/workbook.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common'
 import { RolesModule } from '@libs/auth'
-import { GroupModule } from '@client/group/group.module'
 import { WorkbookAdminController } from './workbook-admin.controller'
 import {
   GroupWorkbookController,
@@ -9,7 +8,7 @@ import {
 import { WorkbookService } from './workbook.service'
 
 @Module({
-  imports: [RolesModule, GroupModule],
+  imports: [RolesModule],
   controllers: [
     WorkbookController,
     GroupWorkbookController,
