+++ backend/libs/auth/src/index.ts
@@ -9,3 +9,5 @@ export * from './jwt/auth-ignore.decorator'
 export * from './roles/roles.module'
 export * from './roles/roles.guard'
 export * from './roles/roles.decorator'
+export * from './roles/group-leader.guard'
+export * from './roles/group-member.guard'
